requirements.txt
langchain==0.2.14
langchain-core==0.2.34
langchain-openai==0.1.21
python-dotenv>=1.0
pydantic>=2.0
faiss-cpu>=1.7.4
tiktoken>=0.3.0
langchain_community
pypdf
azure-ai-formrecognizer>=3.3.3